Integrate bars and phrases
