+++ app/css/index.css
@@ -5,7 +5,6 @@ body {
   font-family: 'Avenir-light', 'Futura', sans-serif;
   height: 100%;
   margin: 0;
-  text-align: center;
   text-transform: uppercase;
 }
