+++ app/negative.html
@@ -23,6 +23,6 @@
     <div class="bar bar6"></div>
   </div>
 
-  <script src="../lib/bars/negative.js"></script>
+  <script src="../lib/main/negative.js"></script>
 </body>
 </html>
