+++ app/positive.html
@@ -23,6 +23,6 @@
     <div class="bar bar6"></div>
   </div>
 
-  <script src="../lib/bars/positive.js"></script>
+  <script src="../lib/main/positive.js"></script>
 </body>
 </html>
