+++ lib/main/negative.js
@@ -1,0 +1,5 @@
+import * as bars from '../bars/negative'
+import * as phraseEngine from '../phrase_engine.js'
+
+bars.animate()
+phraseEngine.run()
