+++ lib/main/positive.js
@@ -1,0 +1,5 @@
+import * as bars from '../bars/positive'
+import * as phraseEngine from '../phrase_engine.js'
+
+bars.animate()
+phraseEngine.run()
