+++ lib/phrase_engine/chooser/solo-renderer-chooser.js
