+++ lib/renderers/apparitionanna.js
@@ -11,7 +11,7 @@ const X_MAX = 1000
 const Y_MIN = 0
 const Y_MAX = 1000
 
-export default class ApparitionRenderer {
+export default class ApparitionAnnaRenderer {
   constructor ({ id: phraseId, text: phraseText }) {
     this.phraseId = phraseId
     this.phraseText = phraseText
