+++ lib/renderers/ticking.js
@@ -7,7 +7,7 @@ const X_MAX = 600
 const Y_MIN = 0
 const Y_MAX = 500
 
-export default class CrazyRenderer {
+export default class TickingRenderer {
   constructor ({ id: phraseId, text: phraseText }) {
     this.phraseId = phraseId
     this.phraseText = phraseText
@@ -30,8 +30,11 @@ export default class CrazyRenderer {
 
     let phraseDiv = window.document.createElement('div')
     phraseDiv.setAttribute('id', `phrase_${this.phraseId}`)
-    phraseDiv.style = `transform:
-      translateX(${baseX0}px) translateY(${baseY0}px);`
+    phraseDiv.style = `
+      mix-blend-mode: difference;
+      transform:
+        translateX(${baseX0}px) translateY(${baseY0}px);
+    `
     let phraseSpan = window.document.createElement('span')
     phraseSpan.textContent = this.phraseText
     phraseDiv.appendChild(phraseSpan)
